+++ lib/features/settings/box_sort_order.dart
@@ -1,0 +1,6 @@
+enum BoxSortOrder {
+  createdOldestFirst,
+  createdNewestFirst,
+  labelAscending,
+  labelDescending,
+}